Move areAllResultsOnPnc function into root of lib folder
And remove areAllResultsOnPnc folder.
