+++ packages/core/phase2/exceptions/HO200101.ts
@@ -5,7 +5,7 @@ import ResultClass from "../../types/ResultClass"
 import checkResultClassExceptions from "./checkResultClassExceptions"
 import areAllPncDisposalsWithType from "../lib/areAllPncDisposalsWithType"
 import { isPncUpdateDataset } from "../../types/PncUpdateDataset"
-import { areAllResultsOnPnc } from "../lib/areAllResultsOnPnc"
+import areAllResultsOnPnc from "../lib/areAllResultsOnPnc"
 import ExceptionCode from "bichard7-next-data-latest/dist/types/ExceptionCode"
 import errorPaths from "../../lib/exceptions/errorPaths"
