+++ packages/core/phase2/exceptions/HO200114.test.ts
@@ -1,7 +1,7 @@
 import ResultClass from "../../types/ResultClass"
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 import HO200114 from "./HO200114"
-import { areAllResultsOnPnc } from "../lib/areAllResultsOnPnc"
+import areAllResultsOnPnc from "../lib/areAllResultsOnPnc"
 import generatePncUpdateDatasetFromOffenceList from "../tests/fixtures/helpers/generatePncUpdateDatasetFromOffenceList"
 
 jest.mock("../lib/areAllResultsOnPnc")
