+++ packages/core/phase2/exceptions/HO200115.test.ts
@@ -2,7 +2,7 @@ import generateAhoFromOffenceList from "../tests/fixtures/helpers/generateAhoFro
 import ResultClass from "../../types/ResultClass"
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 import HO200115 from "./HO200115"
-import { areAllResultsOnPnc } from "../lib/areAllResultsOnPnc"
+import areAllResultsOnPnc from "../lib/areAllResultsOnPnc"
 import type { PncQueryResult } from "../../types/PncQueryResult"
 
 jest.mock("../lib/areAllResultsOnPnc")
