+++ packages/core/phase2/exceptions/HO200124.test.ts
@@ -5,7 +5,7 @@ import type { Offence } from "../../types/AnnotatedHearingOutcome"
 import ResultClass from "../../types/ResultClass"
 import generateFakeAho from "../../phase1/tests/helpers/generateFakeAho"
 import hasUnmatchedPncOffences from "../lib/hasUnmatchedPncOffences"
-import { areAllResultsOnPnc } from "../lib/areAllResultsOnPnc"
+import areAllResultsOnPnc from "../lib/areAllResultsOnPnc"
 import ExceptionCode from "bichard7-next-data-latest/dist/types/ExceptionCode"
 
 jest.mock("../lib/hasUnmatchedPncOffences")
