+++ packages/core/phase2/exceptions/checkClashingCourtCaseOperationsException.test.ts
@@ -4,7 +4,7 @@ import type { Offence } from "../../types/AnnotatedHearingOutcome"
 import checkClashingCourtCaseOperationsException from "./checkClashingCourtCaseOperationsException"
 import { PncOperation } from "../../types/PncOperation"
 import ExceptionCode from "bichard7-next-data-latest/dist/types/ExceptionCode"
-import { areAllResultsOnPnc } from "../lib/areAllResultsOnPnc"
+import areAllResultsOnPnc from "../lib/areAllResultsOnPnc"
 
 jest.mock("../lib/areAllResultsOnPnc")
