--- packages/core/phase2/lib/areAllResultsOnPnc/index.ts
@@ -1,1 +1,0 @@
-export { default as areAllResultsOnPnc } from "./areAllResultsOnPnc"
