--- packages/core/phase2/lib/areAllResultsOnPnc/isMatchToPncAdjudicationAndDisposals/index.ts
@@ -1,1 +1,0 @@
-export { default as isMatchToPncAdjudicationAndDisposals } from "../../isMatchToPncAdjudicationAndDisposals"
