+++ packages/core/phase2/phase2.ts
@@ -12,7 +12,7 @@ import type Phase2Result from "./types/Phase2Result"
 import { Phase2ResultType } from "./types/Phase2Result"
 import generateExceptions from "./exceptions/generateExceptions"
 import { generateOperations } from "./lib/generateOperations"
-import { areAllResultsOnPnc } from "./lib/areAllResultsOnPnc"
+import areAllResultsOnPnc from "./lib/areAllResultsOnPnc"
 
 const phase2 = (inputMessage: AnnotatedHearingOutcome | PncUpdateDataset, auditLogger: AuditLogger): Phase2Result => {
   const correlationId = inputMessage.AnnotatedHearingOutcome.HearingOutcome.Hearing.SourceReference.UniqueID
